--- a/mistralrs-core/src/models/deepseek3.rs
+++ b/mistralrs-core/src/models/deepseek3.rs
@@ -19,7 +19,7 @@ use crate::{
         DeepSeekV2RotaryEmbedding, Mlp, RmsNorm, Sdpa,
     },
     layers_masker::{masked_fill, PastKvLenCache},
-    ops::{BincountOp, NonZeroOp, SplitOp, TopKLastDimOp, TopKOutput},
+    ops::{NonZeroOp, SplitOp, TopKLastDimOp, TopKOutput},
     paged_attention::{AttentionImplementation, ModelConfigMetadata, PagedAttention},
     pipeline::{
         extract_logits,
@@ -29,7 +29,8 @@ use crate::{
     serde_default_fn,
     utils::{progress::NiceProgressBar, unvarbuilder::UnVarBuilder},
 };
-
+use std::collections::HashSet;
+use std::iter::FromIterator;
 serde_default_fn!(f64, routed_scaling_factor, 1.0);
 serde_default_fn!(TopkMethod, topk_method, TopkMethod::Greedy);
 serde_default_fn!(usize, moe_layer_freq, 1);
@@ -640,16 +641,15 @@ impl Moe {
 
     fn moe_infer(&self, xs: &Tensor, topk_ids: &Tensor, topk_weight: &Tensor) -> Result<Tensor> {
         let mut y = xs.zeros_like()?;
-        let counts = topk_ids
-            .flatten_all()?
-            .bincount(self.experts.len() as u32)?;
-        for (i, count) in counts
-            .iter()
-            .enumerate()
-            .take(self.experts_end_idx)
-            .skip(self.experts_start_idx)
-        {
-            if *count == 0 {
+        let topk_weight = if topk_weight.dtype() != xs.dtype() {
+            topk_weight.to_dtype(xs.dtype())?
+        } else {
+            topk_weight.to_owned()
+        };
+        let unique_ids: HashSet<u32> =
+            HashSet::from_iter(topk_ids.to_device(&Device::Cpu)?.flatten_all()?.to_vec1()?);
+        for i in self.experts_start_idx..self.experts_end_idx {
+            if !unique_ids.contains(&(i as u32)) {
                 continue;
             }
             let idx_top = topk_ids.eq(i as f64)?.nonzero()?.t()?;
@@ -667,8 +667,7 @@ impl Moe {
                         .index_select(idx, 0)?
                         .gather(&top.unsqueeze(1)?, 1)?
                         .squeeze(1)?
-                        .unsqueeze(D::Minus1)?
-                        .to_dtype(xs.dtype())?,
+                        .unsqueeze(D::Minus1)?,
                 )?,
                 0,
             )?;